Allow inlined trivial case alternatives
Encode case tables as either:

  + Generic: a base address for a table of templates
  + Specialised: A pair of single atoms (with optional stack pop)

The goal of this is to reduce template code size. Many templates
are required just for encoding trivial case alternatives: a single
atom in a large fixed-size template structure. We shouldn't impact
the run-time of a program, but allow the template code to shrink
considerably by encoding trivial alts inline in the case tables.