Eval primitive ops as postfix, using a primitive stack
This allows us to flatten nested expressions when used as arguments
for primitive operations. With this, we can make longer spinal
applications and avoid the cost of unwinding subexpressions in
arithmetic code.

This is slightly at odds with PRS since we are encouraging expressions
onto the stack, where they cannot benefit from PRS evaluation.
Perhaps one improvement would be to statically check when PRS is
guaranteed to succeed. In this case, we should favour PRS instead of
flattening onto the stack.